+++ src/ask_for_help_bot/handler.py
@@ -452,7 +452,11 @@ class AskForHelpHandler(WenetEventHandler):
         else:
             message_string = message_string.with_text("answer_message_nearby")
 
-        message_string = message_string.with_substitution("question", self.parse_text_with_markdown(emojize(message.question, use_aliases=True))) \
+        """
+        json.loads is used to reconstruct non-ascii characters previously encoded using json.dumps
+        emojize it used to reconstruct emojies previously encoded using demojize
+        """
+        message_string = message_string.with_substitution("question", self.parse_text_with_markdown(emojize(json.loads(message.question), use_aliases=True))) \
             .with_substitution("user", questioning_user.name.first if questioning_user.name.first and not anonymous else self._translator.get_translation_instance(user_object.locale).with_text("anonymous_user").translate()) \
             .translate()
 
@@ -460,7 +464,7 @@ class AskForHelpHandler(WenetEventHandler):
         button_ids = [str(uuid.uuid4()) for _ in range(3)]
         button_data = {
             "task_id": message.task_id,
-            "question": emojize(message.question, use_aliases=True),
+            "question": emojize(json.loads(message.question), use_aliases=True),
             "sensitive": sensitive,
             "username": questioning_user.name.first if questioning_user.name.first and not anonymous else self._translator.get_translation_instance(user_object.locale).with_text("anonymous_user").translate(),
             "related_buttons": button_ids,
@@ -484,7 +488,11 @@ class AskForHelpHandler(WenetEventHandler):
         else:
             message_string = message_string.with_text("answer_message_0")
 
-        message_string = message_string.with_substitution("question", self.parse_text_with_markdown(emojize(message.question, use_aliases=True))) \
+        """
+        json.loads is used to reconstruct non-ascii characters previously encoded using json.dumps
+        emojize it used to reconstruct emojies previously encoded using demojize
+        """
+        message_string = message_string.with_substitution("question", self.parse_text_with_markdown(emojize(json.loads(message.question), use_aliases=True))) \
             .with_substitution("user", questioning_user.name.first if questioning_user.name.first and not anonymous else self._translator.get_translation_instance(user_object.locale).with_text("anonymous_user").translate()) \
             .translate()
 
@@ -492,7 +500,7 @@ class AskForHelpHandler(WenetEventHandler):
         button_ids = [str(uuid.uuid4()) for _ in range(4)]
         button_data = {
             "task_id": message.task_id,
-            "question": emojize(message.question, use_aliases=True),
+            "question": emojize(json.loads(message.question), use_aliases=True),
             "sensitive": sensitive,
             "username": questioning_user.name.first if questioning_user.name.first and not anonymous else self._translator.get_translation_instance(user_object.locale).with_text("anonymous_user").translate(),
             "related_buttons": button_ids,
@@ -509,8 +517,12 @@ class AskForHelpHandler(WenetEventHandler):
         return response
 
     def handle_answered_question(self, message: AnsweredQuestionMessage, user_object: WeNetUserProfile, answerer_user: WeNetUserProfile) -> TelegramRapidAnswerResponse:
-        answer_text = emojize(message.answer, use_aliases=True)
-        question_text = self.parse_text_with_markdown(emojize(message.attributes["question"], use_aliases=True))
+        """
+        json.loads is used to reconstruct non-ascii characters previously encoded using json.dumps
+        emojize it used to reconstruct emojies previously encoded using demojize
+        """
+        answer_text = emojize(json.loads(message.answer), use_aliases=True)
+        question_text = self.parse_text_with_markdown(emojize(json.loads(message.attributes["question"]), use_aliases=True))
         # Translate the message that there is a new answer and insert the details of the question and answer
         message_string = self._translator.get_translation_instance(user_object.locale) \
             .with_text("new_answer_message") \
@@ -539,9 +551,13 @@ class AskForHelpHandler(WenetEventHandler):
 
     def handle_answered_picked(self, message: AnsweredPickedMessage, user_object: WeNetUserProfile) -> TextualResponse:
         # Translate the message that the answer to a question was picked as the best and insert the details of the question
+        """
+        json.loads is used to reconstruct non-ascii characters previously encoded using json.dumps
+        emojize it used to reconstruct emojies previously encoded using demojize
+        """
         message_string = self._translator.get_translation_instance(user_object.locale) \
             .with_text("picked_best_answer") \
-            .with_substitution("question", self.parse_text_with_markdown(emojize(message.attributes["question"], use_aliases=True))) \
+            .with_substitution("question", self.parse_text_with_markdown(emojize(json.loads(message.attributes["question"]), use_aliases=True))) \
             .translate()
         return TextualResponse(message_string)
 
@@ -682,7 +698,11 @@ class AskForHelpHandler(WenetEventHandler):
         user_locale = self._get_user_locale_from_incoming_event(incoming_event)
         response = OutgoingEvent(social_details=incoming_event.social_details)
         if isinstance(incoming_event.incoming_message, IncomingTextMessage):
-            question = demojize(incoming_event.incoming_message.text)
+            """
+            demojize is used handle and encode emojies and them using emojize it is possible to reconstruct them
+            json.dumps is used to handle and encode non-ascii characters and then using json.loads it is possible to reconstruct them
+            """
+            question = json.dumps(demojize(incoming_event.incoming_message.text))
             context = incoming_event.context
             context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_1)
             context.with_static_state(self.CONTEXT_ASKED_QUESTION, question)
@@ -771,11 +791,11 @@ class AskForHelpHandler(WenetEventHandler):
         context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_QUESTION_4)
         context.with_static_state(self.CONTEXT_BELIEF_VALUES_SIMILARITY, intent)
         message = self._translator.get_translation_instance(user_locale).with_text("sensitive_question").translate()
-        button_1_text = self._translator.get_translation_instance(user_locale).with_text("not_sensitive").translate()
-        button_2_text = self._translator.get_translation_instance(user_locale).with_text("sensitive").translate()
+        button_1_text = self._translator.get_translation_instance(user_locale).with_text("sensitive").translate()
+        button_2_text = self._translator.get_translation_instance(user_locale).with_text("not_sensitive").translate()
         response_with_buttons = TelegramRapidAnswerResponse(TextualResponse(message), row_displacement=[1, 1])
-        response_with_buttons.with_textual_option(button_1_text, self.INTENT_NOT_SENSITIVE_QUESTION)
-        response_with_buttons.with_textual_option(button_2_text, self.INTENT_SENSITIVE_QUESTION)
+        response_with_buttons.with_textual_option(button_1_text, self.INTENT_SENSITIVE_QUESTION)
+        response_with_buttons.with_textual_option(button_2_text, self.INTENT_NOT_SENSITIVE_QUESTION)
         response.with_message(response_with_buttons)
         response.with_context(context)
         return response
@@ -970,15 +990,20 @@ class AskForHelpHandler(WenetEventHandler):
         context.delete_static_state(self.CONTEXT_PROPOSED_TASKS)
         user_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
         task = service_api.get_task(button_payload.payload["task_id"])
+
+        """
+        json.loads is used to reconstruct non-ascii characters previously encoded using json.dumps
+        emojize it used to reconstruct emojies previously encoded using demojize
+        """
         if button_payload.payload.get("sensitive", False):
             context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_ANSWERING_SENSITIVE)
             message = self._translator.get_translation_instance(user_locale).with_text("you_are_answering_to_sensitive")\
-                .with_substitution("question", self.parse_text_with_markdown(emojize(task.goal.name, use_aliases=True)))\
+                .with_substitution("question", self.parse_text_with_markdown(emojize(json.loads(task.goal.name), use_aliases=True)))\
                 .translate()
         else:
             context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_ANSWERING)
             message = self._translator.get_translation_instance(user_locale).with_text("you_are_answering_to")\
-                .with_substitution("question", self.parse_text_with_markdown(emojize(task.goal.name, use_aliases=True)))\
+                .with_substitution("question", self.parse_text_with_markdown(emojize(json.loads(task.goal.name), use_aliases=True)))\
                 .translate()
 
         response = OutgoingEvent(social_details=incoming_event.social_details)
@@ -1002,7 +1027,11 @@ class AskForHelpHandler(WenetEventHandler):
         response = OutgoingEvent(social_details=incoming_event.social_details)
         if isinstance(incoming_event.incoming_message, IncomingTextMessage):
             context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_ANSWERING_ANONYMOUSLY)
-            context.with_static_state(self.CONTEXT_ANSWER_TO_QUESTION, demojize(incoming_event.incoming_message.text))
+            """
+            demojize is used handle and encode emojies and them using emojize it is possible to reconstruct them
+            json.dumps is used to handle and encode non-ascii characters and then using json.loads it is possible to reconstruct them
+            """
+            context.with_static_state(self.CONTEXT_ANSWER_TO_QUESTION, json.dumps(demojize(incoming_event.incoming_message.text)))
             message = self._translator.get_translation_instance(user_locale).with_text("answer_anonymously").translate()
             button_1_text = self._translator.get_translation_instance(user_locale).with_text("anonymous_answer_1").translate()
             button_2_text = self._translator.get_translation_instance(user_locale).with_text("anonymous_answer_2").translate()
@@ -1075,7 +1104,11 @@ class AskForHelpHandler(WenetEventHandler):
         response = OutgoingEvent(social_details=incoming_event.social_details)
         if isinstance(incoming_event.incoming_message, IncomingTextMessage):
             question_id = context.get_static_state(self.CONTEXT_QUESTION_TO_ANSWER)
-            answer = demojize(incoming_event.incoming_message.text)
+            """
+            demojize is used handle and encode emojies and them using emojize it is possible to reconstruct them
+            json.dumps is used to handle and encode non-ascii characters and then using json.loads it is possible to reconstruct them
+            """
+            answer = json.dumps(demojize(incoming_event.incoming_message.text))
             actioneer_id = context.get_static_state(self.CONTEXT_WENET_USER_ID)
             try:
                 transaction = TaskTransaction(None, question_id, self.LABEL_ANSWER_TRANSACTION, int(datetime.now().timestamp()), int(datetime.now().timestamp()), actioneer_id, {"answer": answer, "anonymous": False}, [])
@@ -1265,7 +1298,11 @@ class AskForHelpHandler(WenetEventHandler):
         user_locale = self._get_user_locale_from_incoming_event(incoming_event)
         response = OutgoingEvent(social_details=incoming_event.social_details)
         if isinstance(incoming_event.incoming_message, IncomingTextMessage):
-            reason = demojize(incoming_event.incoming_message.text)
+            """
+            demojize is used handle and encode emojies and them using emojize it is possible to reconstruct them
+            json.dumps is used to handle and encode non-ascii characters and then using json.loads it is possible to reconstruct them
+            """
+            reason = json.dumps(demojize(incoming_event.incoming_message.text))
             context = incoming_event.context
             context.with_static_state(self.CONTEXT_CURRENT_STATE, self.STATE_BEST_ANSWER_1)
             context.with_static_state(self.CONTEXT_CHOSEN_ANSWER_REASON, reason)
@@ -1357,7 +1394,11 @@ class AskForHelpHandler(WenetEventHandler):
             for task in tasks:
                 questioning_user = service_api.get_user_profile(str(task.requester_id))
                 if questioning_user:
-                    task_text = f"#{1 + len(proposed_tasks)}: *{self.parse_text_with_markdown(emojize(task.goal.name, use_aliases=True))}* - {questioning_user.name.first if questioning_user.name.first and not task.attributes.get('anonymous') else self._translator.get_translation_instance(user_locale).with_text('anonymous_user').translate()}"
+                    """
+                    json.loads is used to reconstruct non-ascii characters previously encoded using json.dumps
+                    emojize it used to reconstruct emojies previously encoded using demojize
+                    """
+                    task_text = f"#{1 + len(proposed_tasks)}: *{self.parse_text_with_markdown(emojize(json.loads(task.goal.name), use_aliases=True))}* - {questioning_user.name.first if questioning_user.name.first and not task.attributes.get('anonymous') else self._translator.get_translation_instance(user_locale).with_text('anonymous_user').translate()}"
                     if task.attributes.get("sensitive"):
                         task_text = task_text + f" - {self._translator.get_translation_instance(user_locale).with_text('sensitive').translate()}"
                     tasks_texts.append(task_text)